requirements.txt
asyncio
requests
aiohttp
python-dotenv
aiosqlite